Utilização de inner class nas validações de Pedido e Itens de Pedido.
